tty: move WriteStream to internal/tty/write_stream, load node:net eagerly in node:tty
Object.getPrototypeOf(tty.ReadStream) must be net.Socket as soon as node:tty
loads (Node's test-net-access-byteswritten reads it), so the lazy prototype
chain goes. A TTY process.stdout now takes WriteStream from the internal
module and does not load node:net.
